add draggable component
