+++ src/App.vue
@@ -16,7 +16,7 @@ export default {
 </script>
 
 <style>
-/* body {
-  background-color: black
-} */
+body {
+  background-color: darkgrey
+}
 </style>
