+++ src/views/Update.vue
@@ -1,0 +1,12 @@
+<template>
+</template>
+
+<script>
+export default {
+  name: 'Update'
+}
+</script>
+
+<style>
+
+</style>
